requirements.txt
pyyaml==6.0.2
numpy==1.26.4
scikit-learn==0.24.1
opencv-python==4.10.0.84
pillow==11.1.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu117
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.15.0+cu117
xFormers==0.0.18